+++ projects/common/lib/components/phone-number/phone-number.component.ts
@@ -5,21 +5,20 @@ import {
   Output,
   EventEmitter,
   Optional,
-  Self
+  Self,
+  OnInit
 } from '@angular/core';
 import { ControlContainer, NgForm, NgControl } from '@angular/forms';
 import { MaskModel, NUMBER, SPACE } from '../../models/mask.model';
 
-import {  ControlValueAccessor,  NG_VALUE_ACCESSOR } from '@angular/forms';
+import { ControlValueAccessor, NG_VALUE_ACCESSOR } from '@angular/forms';
 /**
  * PhoneNumberComponent is a used to show the Phone number.
  *
- * Make sure to add the direcive `commonValidatePhone`
  *
  * @example
  *         <common-phone-number name='phoneNumber'
  *                              [(ngModel)]='dataService.facAdminPhoneNumber'
- *                              commonValidatePhone
  *                              required></common-phone-number>
  * @export
  */
@@ -31,39 +30,60 @@ import {  ControlValueAccessor,  NG_VALUE_ACCESSOR } from '@angular/forms';
   styleUrls: ['./phone-number.component.scss'],
   /* @deprecate this ViewProvider as we're using ControlValueAccessor @Optional  */
   viewProviders: [
-    { provide: ControlContainer, useExisting: forwardRef(() => NgForm ) }
+    { provide: ControlContainer, useExisting: forwardRef(() => NgForm) }
   ]
 })
 
-export class PhoneNumberComponent extends MaskModel implements ControlValueAccessor {
+export class PhoneNumberComponent extends MaskModel implements ControlValueAccessor, OnInit {
 
   static PhoneNumberRegEx = '^[2-9]{1}\\d{2}[\\-]?\\d{3}[\\-]?\\d{4}$';
   @Input() displayMask: boolean = true;
   @Input() required: boolean = false;
   @Input() label: string = 'Mobile';
 
-  /** @deprecated - use ngModel - rename this to `value`. */
-  @Input() phoneNumber: string = '';
+  @Input() allowInternational: boolean = true;
 
-  /** @deprecate - Do we have any applications that need to use this? */
-  @Input() objectID: string = 'phone_' + this.objectId;
-  @Output() onChange = new EventEmitter<string>();
-  // public mask = ['+','1',SPACE,'(',NUMBER,NUMBER,NUMBER,')',SPACE,NUMBER,NUMBER,NUMBER,'-',NUMBER,NUMBER,NUMBER,NUMBER];
-  public _onChange = (_: any) => {};
-  public _onTouched = () => {};
+  public phoneNumber: string = '';
+
+  public _onChange = (_: any) => { };
+  public _onTouched = () => { };
 
 
-  constructor( @Optional() @Self() public controlDir: NgControl ) {
+  constructor(@Optional() @Self() public controlDir: NgControl) {
     super();
-    if ( controlDir ) {
+    if (controlDir) {
       controlDir.valueAccessor = this;
     }
 
+  }
 
-    this.placeholder = '+1 (555) 555-5555';
-    // Note - we added in the /[2-9]/ regex in order to match MSP's JSON Schema.
-    // Make sure both places match.
-    this.mask = ['+', '1', SPACE, '(', /[2-9]/, NUMBER, NUMBER, ')', SPACE, NUMBER, NUMBER, NUMBER, '-', NUMBER, NUMBER, NUMBER, NUMBER];
+  ngOnInit() {
+    const internationalPrefix = '+1';
+    this.placeholder = '(555) 555-5555';
+    this.mask = ['(', /[2-9]/, NUMBER, NUMBER, ')', SPACE, NUMBER, NUMBER, NUMBER, '-', NUMBER, NUMBER, NUMBER, NUMBER];
+
+    if (this.allowInternational) {
+      this.placeholder = `${internationalPrefix} ${this.placeholder}`;
+      const prefixArrayOfChar = internationalPrefix.split(''); // ['+', '1']
+      this.mask = [...prefixArrayOfChar, SPACE, ...this.mask];
+    }
+
+
+
+    // Register self validation
+    Promise.resolve().then(() => {
+
+      if (this.controlDir) {
+
+        const allValidators = [this.validateSelf.bind(this)];
+        if (this.controlDir.control.validator) {
+          allValidators.push(this.controlDir.control.validator);
+        }
+        this.controlDir.control.setValidators(allValidators);
+        this.controlDir.control.updateValueAndValidity();
+      }
+    });
+
   }
 
   get phoneNumberString(): string {
@@ -72,10 +92,9 @@ export class PhoneNumberComponent extends MaskModel implements ControlValueAcces
 
 
   setPhoneNumber(evt) {
-    console.log('setphonenumber', evt.target.value);
     const value = evt.target.value;
     this.phoneNumber = value;
-    this.onChange.emit(this.phoneNumber);
+    this.valueChange.emit(this.phoneNumber);
     this._onChange(value);
     this._onTouched();
   }
@@ -93,6 +112,29 @@ export class PhoneNumberComponent extends MaskModel implements ControlValueAcces
     // phoneNumber is where the actual data is displayed to user for this
     // component
     this.phoneNumber = value;
+  }
+
+
+  private validateSelf() {
+    const value = this.phoneNumber;
+    const phoneLength = this.allowInternational ? 11 : 10;
+
+    if (value) {
+      const stripped = value
+        .replace(/_/g, '') // remove underlines
+        .replace(/\s/g, '') // spaces
+        .replace(/\+|-/g, '') // + or - symbol
+        .replace('(', '')
+        .replace(')', '');
+
+      const valid = stripped.length === phoneLength;
+
+      console.log('self validation', { valid, stripped, phoneLength });
+      return valid ? null : { required: true };
+
+    }
+    return null;
+
   }
 
 }